fix: Spring 과제 글 수정 삭제 비밀번호 확인 기능 추가

+++ inno_3_spring_basic/Spring_BlogCRUD/src/main/java/com/ldu/spring_blogcrud/service/PostService.java
@@ -47,12 +47,22 @@ public class PostService {
     public Long update(Long id, PostRequestDto postRequestDto) {
         Post post = postRepository.findById(id).orElseThrow(() ->
                 new IllegalArgumentException("게시글이 존재하지 않습니다."));
-        post.update(postRequestDto);
+        if (post.getPassword().equals(postRequestDto.getPassword())) {
+            post.update(postRequestDto);
+        } else {
+            throw new IllegalArgumentException("비밀번호가 틀립니다.");
+        }
         return post.getId();
     }
 
-    public Long delete(Long id) {
-        postRepository.deleteById(id);
+    public Long delete(Long id, PostRequestDto postRequestDto) {
+        Post post = postRepository.findById(id).orElseThrow(() ->
+                new IllegalArgumentException("게시글이 존재하지 않습니다."));
+        if (post.getPassword().equals(postRequestDto.getPassword())) {
+            postRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("비밀번호가 틀립니다.");
+        }
         return id;
     }
